Torch and Lamp
Add lamp that comsumes oil and recover some sanity. I made up some temporary stats to test it.
Add BP_interactable_torch that can be triggered if (1) the interaction function is called and (2) you have a glowing lamp in your hand. (I know it is supposed to be lighten up by matches, but since we dont have that yet, I use the lamp for now.)